requirements.txt
# Required libraries for the project (installs latest versions)
biopython
pandas
matplotlib